fix: editor show header
