+++ components/chat/chat-container.tsx
@@ -1,18 +1,15 @@
 'use client';
-import React, { useContext, useEffect, useMemo, useState } from 'react';
-import { Card, CardContent, Typography, Grid, Skeleton, AspectRatio, Box, aspectRatioClasses } from '@/lib/mui';
+import React, { useContext, useEffect, useState } from 'react';
+import { Card, Grid, Skeleton, AspectRatio, Box, aspectRatioClasses } from '@/lib/mui';
 import { useRequest } from 'ahooks';
 import useAgentChat from '@/hooks/use-agent-chat';
 import Completion from '@/components/chat/completion';
 import { ChartData } from '@/types/chart';
-import LineChart from '../chart/line-chart';
-import BarChart from '../chart/bar-chart';
-import TableChart from '../chart/table-chart';
 import { apiInterceptors, getChatHistory, postChatModeParamsList } from '@/client/api';
 import { ChatContext } from '@/app/chat-context';
 import MuiLoading from '../common/loading';
 import Header from './header';
-import { useSearchParams } from 'next/navigation';
+import Chart from '../chart';
 
 const ChartSkeleton = () => {
   return (
@@ -35,7 +32,6 @@ const ChartSkeleton = () => {
 };
 
 const ChatContainer = () => {
-  const searchParams = useSearchParams();
   const [loading, setLoading] = useState<boolean>(false);
   const [chartsData, setChartsData] = useState<Array<ChartData>>();
   const { refreshDialogList, scene, chatId, model, setModel } = useContext(ChatContext);
@@ -90,36 +86,6 @@ const ChatContainer = () => {
     }
   }, [history]);
 
-  const chartRows = useMemo(() => {
-    if (chartsData) {
-      let res = [];
-      // 若是有类型为 IndicatorValue 的，提出去，独占一行
-      const chartCalc = chartsData?.filter((item) => item.chart_type === 'IndicatorValue');
-      if (chartCalc.length > 0) {
-        res.push({
-          charts: chartCalc,
-          type: 'IndicatorValue',
-        });
-      }
-      let otherCharts = chartsData?.filter((item) => item.chart_type !== 'IndicatorValue');
-      let otherLength = otherCharts.length;
-      let curIndex = 0;
-      // charts 数量 3～8个，暂定每行排序
-      let chartLengthMap = [[0], [1], [2], [1, 2], [1, 3], [2, 1, 2], [2, 1, 3], [3, 1, 3], [3, 2, 3]];
-      chartLengthMap[otherLength].forEach((item) => {
-        if (item > 0) {
-          const rowsItem = otherCharts.slice(curIndex, curIndex + item);
-          curIndex = curIndex + item;
-          res.push({
-            charts: rowsItem,
-          });
-        }
-      });
-      return res;
-    }
-    return undefined;
-  }, [chartsData]);
-
   return (
     <>
       <Header
@@ -130,42 +96,7 @@ const ChatContainer = () => {
       />
       <MuiLoading visible={loading} />
       <div className="px-4 flex flex-1 overflow-hidden">
-        {chartsData && (
-          <div className="w-3/4">
-            <div className="flex flex-col gap-3 h-full">
-              {chartRows?.map((chartRow, index) => (
-                <div key={`chart_row_${index}`} className={`${chartRow?.type !== 'IndicatorValue' ? 'flex gap-3' : ''}`}>
-                  {chartRow.charts.map((chart) => {
-                    if (chart.chart_type === 'IndicatorValue') {
-                      return (
-                        <div key={chart.chart_uid} className="flex flex-row gap-3">
-                          {chart.values.map((item) => (
-                            <div key={item.name} className="flex-1">
-                              <Card sx={{ background: 'transparent' }}>
-                                <CardContent className="justify-around">
-                                  <Typography gutterBottom component="div">
-                                    {item.name}
-                                  </Typography>
-                                  <Typography>{item.value}</Typography>
-                                </CardContent>
-                              </Card>
-                            </div>
-                          ))}
-                        </div>
-                      );
-                    } else if (chart.chart_type === 'LineChart') {
-                      return <LineChart key={chart.chart_uid} chart={chart} />;
-                    } else if (chart.chart_type === 'BarChart') {
-                      return <BarChart key={chart.chart_uid} chart={chart} />;
-                    } else if (chart.chart_type === 'Table') {
-                      return <TableChart key={chart.chart_uid} chart={chart} />;
-                    }
-                  })}
-                </div>
-              ))}
-            </div>
-          </div>
-        )}
+        <Chart chartsData={chartsData} />
         {/** skeleton */}
         {!chartsData && scene === 'chat_dashboard' && (
           <div className="w-3/4 p-6">
