+++ components/chat/db-editor.tsx
@@ -7,9 +7,9 @@ import type { DataNode } from 'antd/es/tree';
 import MonacoEditor from './monaco-editor';
 import { sendGetRequest, sendSpacePostRequest } from '@/utils/request';
 import { useSearchParams } from 'next/navigation';
-import ChartContainer from './chat-container';
 import { OnChange } from '@monaco-editor/react';
 import Header from './header';
+import Chart from '../chart';
 
 const { Search } = Input;
 
@@ -48,14 +48,13 @@ function DbEditorContent({ editorValue, chartData, tableData, handleChange }: IP
     if (!chartData) return <div></div>;
     return (
       <div className="flex-1 overflow-auto p-3" style={{ flexShrink: 0, overflow: 'hidden' }}>
-        <ChartContainer {...chartData} />
+        <Chart chartsData={[chartData]} />
       </div>
     );
   }, [chartData]);
 
   return (
     <>
-      <Header />
       <div className="flex-1 flex overflow-hidden">
         <div className="flex-1" style={{ flexShrink: 0, overflow: 'auto' }}>
           <MonacoEditor value={editorValue?.sql || ''} language="mysql" onChange={handleChange} thoughts={editorValue?.thoughts || ''} />
@@ -418,6 +417,7 @@ function DbEditor() {
 
   return (
     <div className="flex flex-col w-full h-full">
+      <Header />
       <div className="bg-[#f8f8f8] border-[var(--joy-palette-divider)] border-b border-solid flex items-center px-3 justify-between">
         <div className="absolute right-4 top-6">
           <Button
